requirements.txt
kiteconnect
schedule
pandas
requests
python-dotenv
